add documentation and update PEP8 standards
