+++ doubleml/__init__.py
@@ -5,7 +5,8 @@ from .double_ml_pliv import DoubleMLPLIV
 from .double_ml_irm import DoubleMLIRM
 from .double_ml_iivm import DoubleMLIIVM
 from .double_ml_data import DoubleMLData, DoubleMLClusterData
-from .double_ml_qte import DoubleMLPQ, DoubleMLQTE
+from .double_ml_qte import DoubleMLQTE
+from .double_ml_pq import DoubleMLPQ
 
 __all__ = ['DoubleMLPLR',
            'DoubleMLPLIV',
